add basic model concepts to server.js, mostly for testing purposes
